Mobile report budget summary + reuse desktop components
