+++ packages/desktop-client/src/components/budget/report/budgetsummary/ExpenseTotal.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import { reportBudget } from 'loot-core/src/client/queries';
+
+import { type CSSProperties } from '../../../../style';
+
+import BudgetTotal from './BudgetTotal';
+import { ExpenseProgress } from './ExpenseProgress';
+
+type ExpenseTotalProps = {
+  style?: CSSProperties;
+};
+export default function ExpenseTotal({ style }: ExpenseTotalProps) {
+  return (
+    <BudgetTotal
+      title="Expenses"
+      current={reportBudget.totalSpent}
+      target={reportBudget.totalBudgetedExpense}
+      ProgressComponent={ExpenseProgress}
+      style={style}
+    />
+  );
+}
