+++ packages/desktop-client/src/components/budget/report/budgetsummary/fraction.ts
@@ -1,0 +1,10 @@
+export default function fraction(num, denom) {
+  if (denom === 0) {
+    if (num > 0) {
+      return 1;
+    }
+    return 0;
+  }
+
+  return num / denom;
+}
