+++ packages/desktop-client/src/components/budget/rollover/RolloverComponents.tsx
@@ -20,8 +20,6 @@ import { makeAmountGrey } from '../util';
 
 import BalanceTooltip from './BalanceTooltip';
 
-export { BudgetSummary } from './BudgetSummary';
-
 let headerLabelStyle: CSSProperties = {
   flex: 1,
   padding: '0 5px',
@@ -402,3 +400,5 @@ export function IncomeCategoryMonth({
     </View>
   );
 }
+
+export { default as BudgetSummary } from './budgetsummary/BudgetSummary';
