+++ packages/desktop-client/src/components/spreadsheet/NamespaceContext.ts
@@ -1,3 +1,3 @@
 import { createContext } from 'react';
 
-export default createContext(undefined);
+export default createContext<string | undefined>(undefined);
